+++ knowde/feature/auth/api.py
@@ -9,9 +9,9 @@ from fastapi_users import FastAPIUsers
 from knowde.feature.__core__.config import Settings
 from knowde.feature.auth.manager import auth_backend, get_user_manager
 from knowde.feature.auth.schema import UserCreate, UserRead, UserUpdate
-from knowde.primitive.account import User
+from knowde.primitive.user import User
 
-from .sso.route import GOOGLE_URL, router_google_oauth
+from .sso import GOOGLE_URL, router_google_oauth
 
 auth_router = APIRouter()
 rc = FastAPIUsers[User, UUID](get_user_manager, [auth_backend()])
